优先处理矢山
